Sprint6

+++ async_api/auth_service/src/commands.py
@@ -1,31 +1,40 @@
-from sqlalchemy.orm import Session
-from core.config import settings
 import typer
-from typing import Annotated
-from sqlalchemy import create_engine
+from sqlalchemy import create_engine, select
+from sqlalchemy.orm import Session
+
+from core.config import settings
 from models.entity import RoleModel, UserModel
 from services.password_service import PasswordService
 
 app = typer.Typer()
 
 dsn = (
-    f'postgresql://{settings.pg_user}'
+    f'postgresql+psycopg://{settings.pg_user}'
     f':{settings.pg_password}@{settings.pg_host}'
     f':{settings.pg_port}/{settings.pg_name}'
 )
 
-engine = create_engine(dsn, echo=True, future=True)
+
+engine = create_engine(dsn, echo=False, future=True)
 
 
-def create_superadmin_role(role_title: Annotated[str, typer.Argument()] = 'superadmin'):
+def create_role(title: str, description: str):
     with Session(engine) as session:
-        role = RoleModel(title=role_title, description='Superadmin role. All granted')
+        role = session.scalars(select(RoleModel).where(RoleModel.title == title)).one_or_none()
+        if role:
+            return
+
+        role = RoleModel(title=title, description=description)
         session.add(role)
         session.commit()
 
 
 def create_superadmin_user():
     with Session(engine) as session:
+        user = session.scalars(select(UserModel).where(UserModel.login == settings.sa_login)).one_or_none()
+        if user:
+            return
+
         pass_service = PasswordService()
         superadmin = UserModel(
             login=settings.sa_login,
@@ -41,25 +50,34 @@ def create_superadmin_user():
 
 def assign_superadmin_role_to_superadmin():
     with Session(engine) as session:
-        superadmin = session.query(UserModel).filter(UserModel.login == settings.sa_login).first()
-        superadmin_role = session.query(RoleModel).filter(RoleModel.title == 'superadmin').first()
-        superadmin.roles.append(superadmin_role)
+        superadmin_user = session.scalars(select(UserModel).where(UserModel.login == settings.sa_login)).one()
+        superadmin_role = session.scalars(select(RoleModel).where(RoleModel.title == settings.role_super_admin)).one()
+        superadmin_user.roles.append(superadmin_role)
         session.commit()
 
 
 @app.command()
 def create_superadmin():
-    create_superadmin_role()
+    create_role(title=settings.role_super_admin, description='superadmin role description')
     create_superadmin_user()
     assign_superadmin_role_to_superadmin()
 
 
 @app.command()
-def delete_seperadmin():
+def delete_superadmin():
     with Session(engine) as session:
-        user = session.query(UserModel).where(UserModel.login == settings.sa_login).one()
-        session.delete(user)
-        session.commit()
+        user = session.scalars(select(UserModel).where(UserModel.login == settings.sa_login)).one_or_none()
+        if user:
+            session.delete(user)
+            session.commit()
+
+
+@app.command()
+def create_roles():
+    create_role(title=settings.role_admin, description='admin role description')
+    create_role(title=settings.role_subscriber, description='subscriber role description')
+    create_role(title=settings.role_guest, description='guest role description')
+    create_role(title=settings.role_user, description='user role description')
 
 
 if __name__ == '__main__':
